backend/requirements.txt
fastapi
uvicorn
python-jose
langchain
langchain-community
langchain-huggingface
transformers
huggingface-hub
python-dotenv
python-multipart